use utils lib for knack calls
